id changes

+++ src/main/java/com/avensys/rts/contactservice/controller/ContactNewController.java
@@ -7,7 +7,15 @@ import org.springframework.context.MessageSource;
 import org.springframework.context.i18n.LocaleContextHolder;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RestController;
 
 import com.avensys.rts.contactservice.constant.MessageConstants;
 import com.avensys.rts.contactservice.payloadnewrequest.ContactNewRequestDTO;
@@ -22,10 +30,9 @@ import jakarta.validation.Valid;
 @CrossOrigin(origins = "*", allowedHeaders = "*")
 public class ContactNewController {
 
-
-    private final Logger log = LoggerFactory.getLogger(ContactNewController.class);
-    private final ContactNewServiceImpl contactService;
-    private final MessageSource messageSource;
+	private final Logger log = LoggerFactory.getLogger(ContactNewController.class);
+	private final ContactNewServiceImpl contactService;
+	private final MessageSource messageSource;
 
 	@Autowired
 	private JwtUtil jwtUtil;
@@ -57,7 +64,7 @@ public class ContactNewController {
 	}
 
 	@DeleteMapping("/contacts/{id}")
-	public ResponseEntity<Object> deleteContact(@PathVariable Integer id) {
+	public ResponseEntity<Object> deleteContact(@PathVariable Long id) {
 		log.info("Delete contact : Controller ");
 		contactService.deleteContact(id);
 		return ResponseUtil.generateSuccessResponse(null, HttpStatus.OK,
@@ -65,7 +72,7 @@ public class ContactNewController {
 	}
 
 	@PutMapping("/contacts/{id}")
-	public ResponseEntity<Object> updateContact(@PathVariable Integer id,
+	public ResponseEntity<Object> updateContact(@PathVariable Long id,
 			@Valid @RequestBody ContactNewRequestDTO contactRequest,
 			@RequestHeader(name = "Authorization") String token) {
 		log.info("Update contact : Controller ");
