+++ src/main/java/com/avensys/rts/contactservice/repository/ContactNewRepository.java
@@ -1,10 +1,11 @@
 package com.avensys.rts.contactservice.repository;
 
-import com.avensys.rts.contactservice.entity.ContactNewEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.List;
 
-public interface ContactNewRepository extends JpaRepository<ContactNewEntity, Integer> {
-    List<ContactNewEntity> findByEntityTypeAndEntityId(String entityType, Integer entityId);
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.avensys.rts.contactservice.entity.ContactNewEntity;
+
+public interface ContactNewRepository extends JpaRepository<ContactNewEntity, Long> {
+	List<ContactNewEntity> findByEntityTypeAndEntityId(String entityType, Integer entityId);
 }
